early black undo

+++ PoolLeaderboardEngine/src/Killer/KillerGameRow.cs
@@ -3,6 +3,6 @@ namespace PoolLeaderboardEngine.Killer;
 public class KillerGameRow
 {
     public required string PlayerName { get; set; }
-    public byte LivesRemaining { get; set; }
+    public int LivesRemaining { get; set; }
     public bool MissedInSuddenDeath { get; set; } = false;
 }
